Refactor server init to be more testable
