+++ oneapp-server/index.js
@@ -3,44 +3,17 @@ const dotenv = require('dotenv');
 dotenv.config();
 
 const config = require('config');
-const args = require('yargs').argv;
-
-// eslint-disable-next-line import/order
-const logger = require('./logger.config');
-
-logger.info('Starting with environment config: %s', config.util.getEnv('NODE_ENV'));
-
 const express = require('express');
 const expressJwt = require('express-jwt');
-const { ApolloServer } = require('apollo-server-express');
-const { makeExecutableSchema } = require('graphql-tools');
-const { applyMiddleware } = require('graphql-middleware');
-const { constraintDirective, constraintDirectiveTypeDefs } = require('graphql-constraint-directive');
 
 // eslint-disable-next-line no-unused-vars
 const cfConfig = require('./cf.config.js'); // cfConfig must be first import to override config values with CloudFoundary values
-const dataSources = require('./db');
-const AuthenticationService = require('./services/AuthenticationService');
-const { typeDefs, resolvers, permissions } = require('./schema');
-const services = require('./services');
+const logger = require('./logger.config');
+const createServer = require('./server');
 
-// Build the schema
-const schema = makeExecutableSchema({
-  typeDefs: [constraintDirectiveTypeDefs, ...typeDefs],
-  resolvers,
-  schemaTransforms: [constraintDirective()],
-  logger: {
-    log: (e) => logger.error(e),
-  },
-});
+logger.info('Starting with environment config: %s', config.util.getEnv('NODE_ENV'));
 
-// Apply schema middleware
-const schemaWithMiddleware = applyMiddleware(
-  schema,
-  AuthenticationService.buildPermissions(permissions),
-);
-
-// Setup the server
+// Setup the app
 const app = express();
 
 // Configure authentication
@@ -55,43 +28,7 @@ app.use(expressJwt({
   return next(err);
 });
 
-// Setup the GraphQL server
-const server = new ApolloServer({
-  // Specify GraphQL config
-  schema: schemaWithMiddleware,
-
-  // Populate the GraphQL context
-  context: ({ req }) => ({
-    // If logged in, place the auth
-    auth: req.auth || null,
-    // Place the requesting IP (either from load balancer or direct)
-    requestIP: req.headers['x-forwarded-for'] || req.connection.remoteAddress,
-    // Place the language
-    language: services.TranslationService.getLanguage(req.headers['accept-language']),
-    // Place the services
-    services,
-  }),
-
-  // Configure our logger implementation
-  logger,
-  formatError: (error) => {
-    logger.error(error);
-    return {
-      code: error.extensions.code,
-      message: error.message,
-      path: error.path,
-    };
-  },
-
-  // Configure data sources available to GraphQL
-  dataSources: () => dataSources,
-
-  // Debug
-  debug: true,
-
-  // Mocks
-  mocks: args.mock || false,
-});
+const server = createServer();
 server.applyMiddleware({ app, path: config.get('server.path') });
 
 // Start the server
@@ -99,5 +36,3 @@ app.listen({ port: config.get('server.port') }, () => {
   logger.info('Running OneApp GraphQL API server at: %s', `http://localhost:${config.get('server.port')}${server.graphqlPath}`);
   logger.info('Database URL: %s:%d', config.get('database.host'), config.get('database.port'));
 });
-
-module.exports = server;
